Refactor tests to two test files
